+++ src/meta/src/controller/catalog/mod.rs
@@ -14,6 +14,7 @@
 
 mod alter_op;
 mod create_op;
+mod drop_op;
 mod get_op;
 mod list_op;
 mod test;
@@ -202,182 +203,6 @@ impl CatalogController {
 }
 
 impl CatalogController {
-    pub async fn drop_database(
-        &self,
-        database_id: DatabaseId,
-    ) -> MetaResult<(ReleaseContext, NotificationVersion)> {
-        let inner = self.inner.write().await;
-        let txn = inner.db.begin().await?;
-        ensure_object_id(ObjectType::Database, database_id, &txn).await?;
-
-        let streaming_jobs: Vec<ObjectId> = StreamingJob::find()
-            .join(JoinType::InnerJoin, streaming_job::Relation::Object.def())
-            .select_only()
-            .column(streaming_job::Column::JobId)
-            .filter(object::Column::DatabaseId.eq(Some(database_id)))
-            .into_tuple()
-            .all(&txn)
-            .await?;
-
-        let (source_fragments, removed_actors, removed_fragments) =
-            resolve_source_register_info_for_jobs(&txn, streaming_jobs.clone()).await?;
-
-        let state_table_ids: Vec<TableId> = Table::find()
-            .select_only()
-            .column(table::Column::TableId)
-            .filter(
-                table::Column::BelongsToJobId
-                    .is_in(streaming_jobs.clone())
-                    .or(table::Column::TableId.is_in(streaming_jobs.clone())),
-            )
-            .into_tuple()
-            .all(&txn)
-            .await?;
-
-        let source_ids: Vec<SourceId> = Object::find()
-            .select_only()
-            .column(object::Column::Oid)
-            .filter(
-                object::Column::DatabaseId
-                    .eq(Some(database_id))
-                    .and(object::Column::ObjType.eq(ObjectType::Source)),
-            )
-            .into_tuple()
-            .all(&txn)
-            .await?;
-
-        let connections = Connection::find()
-            .inner_join(Object)
-            .filter(object::Column::DatabaseId.eq(Some(database_id)))
-            .all(&txn)
-            .await?
-            .into_iter()
-            .map(|conn| conn.connection_id)
-            .collect_vec();
-
-        // Find affect users with privileges on the database and the objects in the database.
-        let to_update_user_ids: Vec<UserId> = UserPrivilege::find()
-            .select_only()
-            .distinct()
-            .column(user_privilege::Column::UserId)
-            .join(JoinType::InnerJoin, user_privilege::Relation::Object.def())
-            .filter(
-                object::Column::DatabaseId
-                    .eq(database_id)
-                    .or(user_privilege::Column::Oid.eq(database_id)),
-            )
-            .into_tuple()
-            .all(&txn)
-            .await?;
-
-        let fragment_mappings = get_fragment_ids_by_jobs(&txn, streaming_jobs.clone())
-            .await?
-            .into_iter()
-            .map(|fragment_id| PbFragmentWorkerSlotMapping {
-                fragment_id: fragment_id as _,
-                mapping: None,
-            })
-            .collect();
-
-        // The schema and objects in the database will be delete cascade.
-        let res = Object::delete_by_id(database_id).exec(&txn).await?;
-        if res.rows_affected == 0 {
-            return Err(MetaError::catalog_id_not_found("database", database_id));
-        }
-        let user_infos = list_user_info_by_ids(to_update_user_ids, &txn).await?;
-
-        txn.commit().await?;
-
-        self.notify_users_update(user_infos).await;
-        let version = self
-            .notify_frontend(
-                NotificationOperation::Delete,
-                NotificationInfo::Database(PbDatabase {
-                    id: database_id as _,
-                    ..Default::default()
-                }),
-            )
-            .await;
-
-        self.notify_fragment_mapping(NotificationOperation::Delete, fragment_mappings)
-            .await;
-        Ok((
-            ReleaseContext {
-                database_id,
-                streaming_job_ids: streaming_jobs,
-                state_table_ids,
-                source_ids,
-                connections,
-                source_fragments,
-                removed_actors,
-                removed_fragments,
-            },
-            version,
-        ))
-    }
-
-    pub async fn drop_schema(
-        &self,
-        schema_id: SchemaId,
-        drop_mode: DropMode,
-    ) -> MetaResult<NotificationVersion> {
-        let inner = self.inner.write().await;
-        let txn = inner.db.begin().await?;
-        let schema_obj = Object::find_by_id(schema_id)
-            .one(&txn)
-            .await?
-            .ok_or_else(|| MetaError::catalog_id_not_found("schema", schema_id))?;
-        // TODO: support drop schema cascade.
-        if drop_mode == DropMode::Restrict {
-            ensure_schema_empty(schema_id, &txn).await?;
-        } else {
-            return Err(MetaError::permission_denied(
-                "drop schema cascade is not supported yet".to_owned(),
-            ));
-        }
-
-        // Find affect users with privileges on the schema and the objects in the schema.
-        let to_update_user_ids: Vec<UserId> = UserPrivilege::find()
-            .select_only()
-            .distinct()
-            .column(user_privilege::Column::UserId)
-            .join(JoinType::InnerJoin, user_privilege::Relation::Object.def())
-            .filter(
-                object::Column::SchemaId
-                    .eq(schema_id)
-                    .or(user_privilege::Column::Oid.eq(schema_id)),
-            )
-            .into_tuple()
-            .all(&txn)
-            .await?;
-
-        let res = Object::delete(object::ActiveModel {
-            oid: Set(schema_id),
-            ..Default::default()
-        })
-        .exec(&txn)
-        .await?;
-        if res.rows_affected == 0 {
-            return Err(MetaError::catalog_id_not_found("schema", schema_id));
-        }
-        let user_infos = list_user_info_by_ids(to_update_user_ids, &txn).await?;
-
-        txn.commit().await?;
-
-        self.notify_users_update(user_infos).await;
-        let version = self
-            .notify_frontend(
-                NotificationOperation::Delete,
-                NotificationInfo::Schema(PbSchema {
-                    id: schema_id as _,
-                    database_id: schema_obj.database_id.unwrap() as _,
-                    ..Default::default()
-                }),
-            )
-            .await;
-        Ok(version)
-    }
-
     pub async fn finish_create_subscription_catalog(&self, subscription_id: u32) -> MetaResult<()> {
         let inner = self.inner.write().await;
         let txn = inner.db.begin().await?;
@@ -470,12 +295,6 @@ impl CatalogController {
             .await?;
         txn.commit().await?;
         Ok(())
-    }
-
-    pub async fn has_any_streaming_jobs(&self) -> MetaResult<bool> {
-        let inner = self.inner.read().await;
-        let count = streaming_job::Entity::find().count(&inner.db).await?;
-        Ok(count > 0)
     }
 
     /// `clean_dirty_creating_jobs` cleans up creating jobs that are creating in Foreground mode or in Initial status.
@@ -619,402 +438,6 @@ impl CatalogController {
         Ok(associated_source_ids)
     }
 
-    async fn log_cleaned_dirty_jobs(
-        &self,
-        dirty_objs: &[PartialObject],
-        txn: &DatabaseTransaction,
-    ) -> MetaResult<()> {
-        // Record cleaned streaming jobs in event logs.
-        let mut dirty_table_ids = vec![];
-        let mut dirty_source_ids = vec![];
-        let mut dirty_sink_ids = vec![];
-        for dirty_job_obj in dirty_objs {
-            let job_id = dirty_job_obj.oid;
-            let job_type = dirty_job_obj.obj_type;
-            match job_type {
-                ObjectType::Table | ObjectType::Index => dirty_table_ids.push(job_id),
-                ObjectType::Source => dirty_source_ids.push(job_id),
-                ObjectType::Sink => dirty_sink_ids.push(job_id),
-                _ => unreachable!("unexpected streaming job type"),
-            }
-        }
-
-        let mut event_logs = vec![];
-        if !dirty_table_ids.is_empty() {
-            let table_info: Vec<(TableId, String, String)> = Table::find()
-                .select_only()
-                .columns([
-                    table::Column::TableId,
-                    table::Column::Name,
-                    table::Column::Definition,
-                ])
-                .filter(table::Column::TableId.is_in(dirty_table_ids))
-                .into_tuple()
-                .all(txn)
-                .await?;
-            for (table_id, name, definition) in table_info {
-                let event = risingwave_pb::meta::event_log::EventDirtyStreamJobClear {
-                    id: table_id as _,
-                    name,
-                    definition,
-                    error: "clear during recovery".to_owned(),
-                };
-                event_logs.push(risingwave_pb::meta::event_log::Event::DirtyStreamJobClear(
-                    event,
-                ));
-            }
-        }
-        if !dirty_source_ids.is_empty() {
-            let source_info: Vec<(SourceId, String, String)> = Source::find()
-                .select_only()
-                .columns([
-                    source::Column::SourceId,
-                    source::Column::Name,
-                    source::Column::Definition,
-                ])
-                .filter(source::Column::SourceId.is_in(dirty_source_ids))
-                .into_tuple()
-                .all(txn)
-                .await?;
-            for (source_id, name, definition) in source_info {
-                let event = risingwave_pb::meta::event_log::EventDirtyStreamJobClear {
-                    id: source_id as _,
-                    name,
-                    definition,
-                    error: "clear during recovery".to_owned(),
-                };
-                event_logs.push(risingwave_pb::meta::event_log::Event::DirtyStreamJobClear(
-                    event,
-                ));
-            }
-        }
-        if !dirty_sink_ids.is_empty() {
-            let sink_info: Vec<(SinkId, String, String)> = Sink::find()
-                .select_only()
-                .columns([
-                    sink::Column::SinkId,
-                    sink::Column::Name,
-                    sink::Column::Definition,
-                ])
-                .filter(sink::Column::SinkId.is_in(dirty_sink_ids))
-                .into_tuple()
-                .all(txn)
-                .await?;
-            for (sink_id, name, definition) in sink_info {
-                let event = risingwave_pb::meta::event_log::EventDirtyStreamJobClear {
-                    id: sink_id as _,
-                    name,
-                    definition,
-                    error: "clear during recovery".to_owned(),
-                };
-                event_logs.push(risingwave_pb::meta::event_log::Event::DirtyStreamJobClear(
-                    event,
-                ));
-            }
-        }
-        self.env.event_log_manager_ref().add_event_logs(event_logs);
-        Ok(())
-    }
-
-    async fn clean_dirty_sink_downstreams(txn: &DatabaseTransaction) -> MetaResult<bool> {
-        // clean incoming sink from (table)
-        // clean upstream fragment ids from (fragment)
-        // clean stream node from (fragment)
-        // clean upstream actor ids from (actor)
-        let all_fragment_ids: Vec<FragmentId> = Fragment::find()
-            .select_only()
-            .columns(vec![fragment::Column::FragmentId])
-            .into_tuple()
-            .all(txn)
-            .await?;
-
-        let all_fragment_ids: HashSet<_> = all_fragment_ids.into_iter().collect();
-
-        let table_sink_ids: Vec<ObjectId> = Sink::find()
-            .select_only()
-            .column(sink::Column::SinkId)
-            .filter(sink::Column::TargetTable.is_not_null())
-            .into_tuple()
-            .all(txn)
-            .await?;
-
-        let all_table_with_incoming_sinks: Vec<(ObjectId, I32Array)> = Table::find()
-            .select_only()
-            .columns(vec![table::Column::TableId, table::Column::IncomingSinks])
-            .into_tuple()
-            .all(txn)
-            .await?;
-
-        let table_incoming_sinks_to_update = all_table_with_incoming_sinks
-            .into_iter()
-            .filter(|(_, incoming_sinks)| {
-                let inner_ref = incoming_sinks.inner_ref();
-                !inner_ref.is_empty()
-                    && inner_ref
-                        .iter()
-                        .any(|sink_id| !table_sink_ids.contains(sink_id))
-            })
-            .collect_vec();
-
-        let new_table_incoming_sinks = table_incoming_sinks_to_update
-            .into_iter()
-            .map(|(table_id, incoming_sinks)| {
-                let new_incoming_sinks = incoming_sinks
-                    .into_inner()
-                    .extract_if(|id| table_sink_ids.contains(id))
-                    .collect_vec();
-                (table_id, I32Array::from(new_incoming_sinks))
-            })
-            .collect_vec();
-
-        // no need to update, returning
-        if new_table_incoming_sinks.is_empty() {
-            return Ok(false);
-        }
-
-        for (table_id, new_incoming_sinks) in new_table_incoming_sinks {
-            tracing::info!("cleaning dirty table sink downstream table {}", table_id);
-            Table::update_many()
-                .col_expr(table::Column::IncomingSinks, new_incoming_sinks.into())
-                .filter(table::Column::TableId.eq(table_id))
-                .exec(txn)
-                .await?;
-
-            let fragments: Vec<(FragmentId, I32Array, StreamNode, i32)> = Fragment::find()
-                .select_only()
-                .columns(vec![
-                    fragment::Column::FragmentId,
-                    fragment::Column::UpstreamFragmentId,
-                    fragment::Column::StreamNode,
-                    fragment::Column::FragmentTypeMask,
-                ])
-                .filter(fragment::Column::JobId.eq(table_id))
-                .into_tuple()
-                .all(txn)
-                .await?;
-
-            for (fragment_id, upstream_fragment_ids, stream_node, fragment_mask) in fragments {
-                let mut upstream_fragment_ids = upstream_fragment_ids.into_inner();
-
-                let dirty_upstream_fragment_ids = upstream_fragment_ids
-                    .extract_if(|id| !all_fragment_ids.contains(id))
-                    .collect_vec();
-
-                if !dirty_upstream_fragment_ids.is_empty() {
-                    // dirty downstream should be materialize fragment of table
-                    assert!(fragment_mask & FragmentTypeFlag::Mview as i32 > 0);
-
-                    tracing::info!(
-                        "cleaning dirty table sink fragment {:?} from downstream fragment {}",
-                        dirty_upstream_fragment_ids,
-                        fragment_id
-                    );
-
-                    let mut pb_stream_node = stream_node.to_protobuf();
-
-                    visit_stream_node_cont_mut(&mut pb_stream_node, |node| {
-                        if let Some(NodeBody::Union(_)) = node.node_body {
-                            node.input.retain_mut(|input| {
-                                if let Some(NodeBody::Merge(merge_node)) = &mut input.node_body
-                                    && all_fragment_ids
-                                        .contains(&(merge_node.upstream_fragment_id as i32))
-                                {
-                                    true
-                                } else {
-                                    false
-                                }
-                            });
-                        }
-                        true
-                    });
-
-                    Fragment::update_many()
-                        .col_expr(
-                            fragment::Column::UpstreamFragmentId,
-                            I32Array::from(upstream_fragment_ids).into(),
-                        )
-                        .col_expr(
-                            fragment::Column::StreamNode,
-                            StreamNode::from(&pb_stream_node).into(),
-                        )
-                        .filter(fragment::Column::FragmentId.eq(fragment_id))
-                        .exec(txn)
-                        .await?;
-
-                    let actors: Vec<(ActorId, ActorUpstreamActors)> = Actor::find()
-                        .select_only()
-                        .columns(vec![
-                            actor::Column::ActorId,
-                            actor::Column::UpstreamActorIds,
-                        ])
-                        .filter(actor::Column::FragmentId.eq(fragment_id))
-                        .into_tuple()
-                        .all(txn)
-                        .await?;
-
-                    for (actor_id, upstream_actor_ids) in actors {
-                        let mut upstream_actor_ids = upstream_actor_ids.into_inner();
-
-                        let dirty_actor_upstreams = upstream_actor_ids
-                            .extract_if(|id, _| !all_fragment_ids.contains(id))
-                            .map(|(id, _)| id)
-                            .collect_vec();
-
-                        if !dirty_actor_upstreams.is_empty() {
-                            tracing::debug!(
-                                "cleaning dirty table sink fragment {:?} from downstream fragment {} actor {}",
-                                dirty_actor_upstreams,
-                                fragment_id,
-                                actor_id,
-                            );
-
-                            Actor::update_many()
-                                .col_expr(
-                                    actor::Column::UpstreamActorIds,
-                                    ActorUpstreamActors::from(upstream_actor_ids).into(),
-                                )
-                                .filter(actor::Column::ActorId.eq(actor_id))
-                                .exec(txn)
-                                .await?;
-                        }
-                    }
-                }
-            }
-        }
-
-        Ok(true)
-    }
-
-    pub async fn drop_function(&self, function_id: FunctionId) -> MetaResult<NotificationVersion> {
-        let inner = self.inner.write().await;
-        let txn = inner.db.begin().await?;
-        let function_obj = Object::find_by_id(function_id)
-            .one(&txn)
-            .await?
-            .ok_or_else(|| MetaError::catalog_id_not_found("function", function_id))?;
-        ensure_object_not_refer(ObjectType::Function, function_id, &txn).await?;
-
-        // Find affect users with privileges on the function.
-        let to_update_user_ids: Vec<UserId> = UserPrivilege::find()
-            .select_only()
-            .distinct()
-            .column(user_privilege::Column::UserId)
-            .filter(user_privilege::Column::Oid.eq(function_id))
-            .into_tuple()
-            .all(&txn)
-            .await?;
-
-        let res = Object::delete_by_id(function_id).exec(&txn).await?;
-        if res.rows_affected == 0 {
-            return Err(MetaError::catalog_id_not_found("function", function_id));
-        }
-        let user_infos = list_user_info_by_ids(to_update_user_ids, &txn).await?;
-
-        txn.commit().await?;
-
-        self.notify_users_update(user_infos).await;
-        let version = self
-            .notify_frontend(
-                NotificationOperation::Delete,
-                NotificationInfo::Function(PbFunction {
-                    id: function_id as _,
-                    schema_id: function_obj.schema_id.unwrap() as _,
-                    database_id: function_obj.database_id.unwrap() as _,
-                    ..Default::default()
-                }),
-            )
-            .await;
-        Ok(version)
-    }
-
-    pub async fn drop_secret(&self, secret_id: SecretId) -> MetaResult<NotificationVersion> {
-        let inner = self.inner.write().await;
-        let txn = inner.db.begin().await?;
-        let (secret, secret_obj) = Secret::find_by_id(secret_id)
-            .find_also_related(Object)
-            .one(&txn)
-            .await?
-            .ok_or_else(|| MetaError::catalog_id_not_found("secret", secret_id))?;
-        ensure_object_not_refer(ObjectType::Secret, secret_id, &txn).await?;
-
-        // Find affect users with privileges on the secret.
-        let to_update_user_ids: Vec<UserId> = UserPrivilege::find()
-            .select_only()
-            .distinct()
-            .column(user_privilege::Column::UserId)
-            .filter(user_privilege::Column::Oid.eq(secret_id))
-            .into_tuple()
-            .all(&txn)
-            .await?;
-
-        let res = Object::delete_by_id(secret_id).exec(&txn).await?;
-        if res.rows_affected == 0 {
-            return Err(MetaError::catalog_id_not_found("secret", secret_id));
-        }
-        let user_infos = list_user_info_by_ids(to_update_user_ids, &txn).await?;
-
-        txn.commit().await?;
-
-        let pb_secret: PbSecret = ObjectModel(secret, secret_obj.unwrap()).into();
-
-        self.notify_users_update(user_infos).await;
-
-        LocalSecretManager::global().remove_secret(pb_secret.id);
-        self.env
-            .notification_manager()
-            .notify_compute_without_version(Operation::Delete, Info::Secret(pb_secret.clone()));
-        let version = self
-            .notify_frontend(
-                NotificationOperation::Delete,
-                NotificationInfo::Secret(pb_secret),
-            )
-            .await;
-        Ok(version)
-    }
-
-    pub async fn drop_connection(
-        &self,
-        connection_id: ConnectionId,
-    ) -> MetaResult<(NotificationVersion, PbConnection)> {
-        let inner = self.inner.write().await;
-        let txn = inner.db.begin().await?;
-        let (conn, conn_obj) = Connection::find_by_id(connection_id)
-            .find_also_related(Object)
-            .one(&txn)
-            .await?
-            .ok_or_else(|| MetaError::catalog_id_not_found("connection", connection_id))?;
-        ensure_object_not_refer(ObjectType::Connection, connection_id, &txn).await?;
-
-        // Find affect users with privileges on the connection.
-        let to_update_user_ids: Vec<UserId> = UserPrivilege::find()
-            .select_only()
-            .distinct()
-            .column(user_privilege::Column::UserId)
-            .filter(user_privilege::Column::Oid.eq(connection_id))
-            .into_tuple()
-            .all(&txn)
-            .await?;
-
-        let res = Object::delete_by_id(connection_id).exec(&txn).await?;
-        if res.rows_affected == 0 {
-            return Err(MetaError::catalog_id_not_found("connection", connection_id));
-        }
-        let user_infos = list_user_info_by_ids(to_update_user_ids, &txn).await?;
-
-        txn.commit().await?;
-
-        let pb_connection: PbConnection = ObjectModel(conn, conn_obj.unwrap()).into();
-
-        self.notify_users_update(user_infos).await;
-        let version = self
-            .notify_frontend(
-                NotificationOperation::Delete,
-                NotificationInfo::Connection(pb_connection.clone()),
-            )
-            .await;
-        Ok((version, pb_connection))
-    }
-
     pub async fn comment_on(&self, comment: PbComment) -> MetaResult<NotificationVersion> {
         let inner = self.inner.write().await;
         let txn = inner.db.begin().await?;
@@ -1072,325 +495,6 @@ impl CatalogController {
             .await;
 
         Ok(version)
-    }
-
-    pub async fn drop_relation(
-        &self,
-        object_type: ObjectType,
-        object_id: ObjectId,
-        drop_mode: DropMode,
-    ) -> MetaResult<(ReleaseContext, NotificationVersion)> {
-        let inner = self.inner.write().await;
-        let txn = inner.db.begin().await?;
-        let obj: PartialObject = Object::find_by_id(object_id)
-            .into_partial_model()
-            .one(&txn)
-            .await?
-            .ok_or_else(|| MetaError::catalog_id_not_found(object_type.as_str(), object_id))?;
-        assert_eq!(obj.obj_type, object_type);
-        let database_id = obj
-            .database_id
-            .ok_or_else(|| anyhow!("dropped object should have database_id"))?;
-
-        let mut to_drop_objects = match drop_mode {
-            DropMode::Cascade => get_referring_objects_cascade(object_id, &txn).await?,
-            DropMode::Restrict => {
-                ensure_object_not_refer(object_type, object_id, &txn).await?;
-                if obj.obj_type == ObjectType::Table {
-                    let indexes = get_referring_objects(object_id, &txn).await?;
-                    assert!(
-                        indexes.iter().all(|obj| obj.obj_type == ObjectType::Index),
-                        "only index could be dropped in restrict mode"
-                    );
-                    indexes
-                } else {
-                    vec![]
-                }
-            }
-        };
-        assert!(
-            to_drop_objects.iter().all(|obj| matches!(
-                obj.obj_type,
-                ObjectType::Table
-                    | ObjectType::Index
-                    | ObjectType::Sink
-                    | ObjectType::View
-                    | ObjectType::Subscription
-            )),
-            "only these objects will depends on others"
-        );
-        to_drop_objects.push(obj);
-
-        // Special handling for 'sink into table'.
-        if object_type != ObjectType::Sink {
-            // When dropping a table downstream, all incoming sinks of the table should be dropped as well.
-            if object_type == ObjectType::Table {
-                let table = Table::find_by_id(object_id)
-                    .one(&txn)
-                    .await?
-                    .ok_or_else(|| MetaError::catalog_id_not_found("table", object_id))?;
-
-                let incoming_sinks = table.incoming_sinks.into_inner();
-
-                if !incoming_sinks.is_empty() {
-                    let objs: Vec<PartialObject> = Object::find()
-                        .filter(object::Column::Oid.is_in(incoming_sinks))
-                        .into_partial_model()
-                        .all(&txn)
-                        .await?;
-
-                    to_drop_objects.extend(objs);
-                }
-            }
-
-            let to_drop_object_ids: HashSet<_> =
-                to_drop_objects.iter().map(|obj| obj.oid).collect();
-
-            // When there is a table sink in the dependency chain of drop cascade, an error message needs to be returned currently to manually drop the sink.
-            for obj in &to_drop_objects {
-                if obj.obj_type == ObjectType::Sink {
-                    let sink = Sink::find_by_id(obj.oid)
-                        .one(&txn)
-                        .await?
-                        .ok_or_else(|| MetaError::catalog_id_not_found("sink", obj.oid))?;
-
-                    // Since dropping the sink into the table requires the frontend to handle some of the logic (regenerating the plan), it’s not compatible with the current cascade dropping.
-                    if let Some(target_table) = sink.target_table
-                        && !to_drop_object_ids.contains(&target_table)
-                    {
-                        bail!(
-                            "Found sink into table with sink id {} in dependency, please drop them manually",
-                            obj.oid,
-                        );
-                    }
-                }
-            }
-        }
-
-        let to_drop_table_ids = to_drop_objects
-            .iter()
-            .filter(|obj| obj.obj_type == ObjectType::Table || obj.obj_type == ObjectType::Index)
-            .map(|obj| obj.oid);
-        let mut to_drop_streaming_jobs = to_drop_objects
-            .iter()
-            .filter(|obj| {
-                obj.obj_type == ObjectType::Table
-                    || obj.obj_type == ObjectType::Sink
-                    || obj.obj_type == ObjectType::Index
-            })
-            .map(|obj| obj.oid)
-            .collect_vec();
-
-        // source streaming job.
-        if object_type == ObjectType::Source {
-            let source_info: Option<StreamSourceInfo> = Source::find_by_id(object_id)
-                .select_only()
-                .column(source::Column::SourceInfo)
-                .into_tuple()
-                .one(&txn)
-                .await?
-                .ok_or_else(|| MetaError::catalog_id_not_found("source", object_id))?;
-            if let Some(source_info) = source_info
-                && source_info.to_protobuf().is_shared()
-            {
-                to_drop_streaming_jobs.push(object_id);
-            }
-        }
-
-        let creating = StreamingJob::find()
-            .filter(
-                streaming_job::Column::JobStatus
-                    .ne(JobStatus::Created)
-                    .and(streaming_job::Column::JobId.is_in(to_drop_streaming_jobs.clone())),
-            )
-            .count(&txn)
-            .await?;
-        if creating != 0 {
-            return Err(MetaError::permission_denied(format!(
-                "can not drop {creating} creating streaming job, please cancel them firstly"
-            )));
-        }
-
-        let mut to_drop_state_table_ids = to_drop_table_ids.clone().collect_vec();
-
-        // Add associated sources.
-        let mut to_drop_source_ids: Vec<SourceId> = Table::find()
-            .select_only()
-            .column(table::Column::OptionalAssociatedSourceId)
-            .filter(
-                table::Column::TableId
-                    .is_in(to_drop_table_ids)
-                    .and(table::Column::OptionalAssociatedSourceId.is_not_null()),
-            )
-            .into_tuple()
-            .all(&txn)
-            .await?;
-        let to_drop_source_objs: Vec<PartialObject> = Object::find()
-            .filter(object::Column::Oid.is_in(to_drop_source_ids.clone()))
-            .into_partial_model()
-            .all(&txn)
-            .await?;
-        to_drop_objects.extend(to_drop_source_objs);
-        if object_type == ObjectType::Source {
-            to_drop_source_ids.push(object_id);
-        }
-
-        if !to_drop_streaming_jobs.is_empty() {
-            let to_drop_internal_table_objs: Vec<PartialObject> = Object::find()
-                .select_only()
-                .columns([
-                    object::Column::Oid,
-                    object::Column::ObjType,
-                    object::Column::SchemaId,
-                    object::Column::DatabaseId,
-                ])
-                .join(JoinType::InnerJoin, object::Relation::Table.def())
-                .filter(table::Column::BelongsToJobId.is_in(to_drop_streaming_jobs.clone()))
-                .into_partial_model()
-                .all(&txn)
-                .await?;
-
-            to_drop_state_table_ids.extend(to_drop_internal_table_objs.iter().map(|obj| obj.oid));
-            to_drop_objects.extend(to_drop_internal_table_objs);
-        }
-
-        let to_drop_objects = to_drop_objects;
-
-        to_drop_objects.iter().for_each(|obj| {
-            if let Some(obj_database_id) = obj.database_id {
-                assert_eq!(
-                    database_id, obj_database_id,
-                    "dropped objects not in the same database: {:?}",
-                    obj
-                );
-            }
-        });
-
-        let (source_fragments, removed_actors, removed_fragments) =
-            resolve_source_register_info_for_jobs(&txn, to_drop_streaming_jobs.clone()).await?;
-
-        let fragment_ids = get_fragment_ids_by_jobs(&txn, to_drop_streaming_jobs.clone()).await?;
-
-        // Find affect users with privileges on all this objects.
-        let to_update_user_ids: Vec<UserId> = UserPrivilege::find()
-            .select_only()
-            .distinct()
-            .column(user_privilege::Column::UserId)
-            .filter(user_privilege::Column::Oid.is_in(to_drop_objects.iter().map(|obj| obj.oid)))
-            .into_tuple()
-            .all(&txn)
-            .await?;
-
-        // delete all in to_drop_objects.
-        let res = Object::delete_many()
-            .filter(object::Column::Oid.is_in(to_drop_objects.iter().map(|obj| obj.oid)))
-            .exec(&txn)
-            .await?;
-        if res.rows_affected == 0 {
-            return Err(MetaError::catalog_id_not_found(
-                object_type.as_str(),
-                object_id,
-            ));
-        }
-        let user_infos = list_user_info_by_ids(to_update_user_ids, &txn).await?;
-
-        txn.commit().await?;
-
-        // notify about them.
-        self.notify_users_update(user_infos).await;
-        let relation_group = build_relation_group_for_delete(to_drop_objects);
-
-        let version = self
-            .notify_frontend(NotificationOperation::Delete, relation_group)
-            .await;
-
-        let fragment_mappings = fragment_ids
-            .into_iter()
-            .map(|fragment_id| PbFragmentWorkerSlotMapping {
-                fragment_id: fragment_id as _,
-                mapping: None,
-            })
-            .collect();
-
-        self.notify_fragment_mapping(NotificationOperation::Delete, fragment_mappings)
-            .await;
-
-        Ok((
-            ReleaseContext {
-                database_id,
-                streaming_job_ids: to_drop_streaming_jobs,
-                state_table_ids: to_drop_state_table_ids,
-                source_ids: to_drop_source_ids,
-                connections: vec![],
-                source_fragments,
-                removed_actors,
-                removed_fragments,
-            },
-            version,
-        ))
-    }
-
-    pub async fn find_creating_streaming_job_ids(
-        &self,
-        infos: Vec<PbCreatingJobInfo>,
-    ) -> MetaResult<Vec<ObjectId>> {
-        let inner = self.inner.read().await;
-
-        type JobKey = (DatabaseId, SchemaId, String);
-
-        // Index table is already included if we still assign the same name for index table as the index.
-        let creating_tables: Vec<(ObjectId, String, DatabaseId, SchemaId)> = Table::find()
-            .select_only()
-            .columns([table::Column::TableId, table::Column::Name])
-            .columns([object::Column::DatabaseId, object::Column::SchemaId])
-            .join(JoinType::InnerJoin, table::Relation::Object1.def())
-            .join(JoinType::InnerJoin, object::Relation::StreamingJob.def())
-            .filter(streaming_job::Column::JobStatus.eq(JobStatus::Creating))
-            .into_tuple()
-            .all(&inner.db)
-            .await?;
-        let creating_sinks: Vec<(ObjectId, String, DatabaseId, SchemaId)> = Sink::find()
-            .select_only()
-            .columns([sink::Column::SinkId, sink::Column::Name])
-            .columns([object::Column::DatabaseId, object::Column::SchemaId])
-            .join(JoinType::InnerJoin, sink::Relation::Object.def())
-            .join(JoinType::InnerJoin, object::Relation::StreamingJob.def())
-            .filter(streaming_job::Column::JobStatus.eq(JobStatus::Creating))
-            .into_tuple()
-            .all(&inner.db)
-            .await?;
-        let creating_subscriptions: Vec<(ObjectId, String, DatabaseId, SchemaId)> =
-            Subscription::find()
-                .select_only()
-                .columns([
-                    subscription::Column::SubscriptionId,
-                    subscription::Column::Name,
-                ])
-                .columns([object::Column::DatabaseId, object::Column::SchemaId])
-                .join(JoinType::InnerJoin, subscription::Relation::Object.def())
-                .join(JoinType::InnerJoin, object::Relation::StreamingJob.def())
-                .filter(streaming_job::Column::JobStatus.eq(JobStatus::Creating))
-                .into_tuple()
-                .all(&inner.db)
-                .await?;
-
-        let mut job_mapping: HashMap<JobKey, ObjectId> = creating_tables
-            .into_iter()
-            .chain(creating_sinks.into_iter())
-            .chain(creating_subscriptions.into_iter())
-            .map(|(id, name, database_id, schema_id)| ((database_id, schema_id, name), id))
-            .collect();
-
-        Ok(infos
-            .into_iter()
-            .flat_map(|info| {
-                job_mapping.remove(&(
-                    info.database_id as _,
-                    info.schema_id as _,
-                    info.name.clone(),
-                ))
-            })
-            .collect())
     }
 }
